First revision
Changed some typos and formatting errors. Formatting errors still exist when math mode text is included inside asterisks for italics, because I'm not sure that adding a bunch of additional asterisks to cut out the math mode portions is the proper course of action.

Tried to make code in certain examples a little cleaner/transparent. Left a mention of the zee() command, but converted most of the code to use the .aut() method on partitions, since it doesn't require extra steps of importing an alias from sfa.

Proposed some changes to the q,t Catalan exercise.

There was one example with q-binomial coefficients where I have no idea what exactly was supposed to be happening.

There's a brief explanation about how one can think of X as being 1 # s[1] and Y as being s[1] # 1 which appears a section or two after that trick is first used, and should be moved to the beginning of the hopf algebra section.